rename bar chart to match update
